requirements.txt
pandas==2.2.3
seaborn==0.13.2
pyyaml==6.0.2